refactor(cli): refactor the TensorBay "CLI" as a module
Restruct contents as following:
1. Use "import outside toplevel" methods to implement and restruct
command-line framework.
2. Separate "cli" as a module in SDK.

PR Closed: https://github.com/Graviti-AI/tensorbay-python-sdk/pull/479
