--- docs/source/api/client/cli.rst
@@ -1,6 +1,0 @@
-tensorbay.client.cli
-====================
-
-.. automodule:: tensorbay.client.cli
-   :members:
-   :show-inheritance:
