+++ docs/source/api/client/client_module.rst
@@ -4,7 +4,6 @@ tensorbay.client
 .. toctree::
    :maxdepth: 4
 
-   cli
    dataset
    exceptions
    gas
