+++ setup.cfg
@@ -54,4 +54,4 @@ exclude = *.tests
 tensorbay = opendataset/*/catalog*.json
 
 [options.entry_points]
-console_scripts = gas = tensorbay.client.cli:cli
+console_scripts = gas = tensorbay.cli.cli:cli
