+++ tensorbay/cli/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+#
+# Copyright 2021 Graviti. All Rights Reserved.
+#
+
+"""Graviti TensorBay CLI."""
